refactor(csharp): merge all C# source files into a single CsmMassData.cs (#45)
Agent-Logs-Url: https://github.com/NEVSTOP-LAB/CSM-MassData-Parameter-Support/sessions/34a505b3-a9e5-4b45-892d-01d7809dc6cb
